217B Faza 2: nowy pakiet services/multi_runtime, shims gemma4_audio_runtime
Tworzy kanoniczny pakiet services/multi_runtime/ (engine, main, schemas,
audio) jako cel migracji; redukuje services/gemma4_audio_runtime/ do
czystych shim re-exportów — żaden kod zewnętrzny nie psuje się podczas
przejścia. Aktualizuje testy i skrypt startowy do nowych ścieżek importu.
Dodaje skrypty audytu i strażnika migracji (8/8 warunków spełnionych).
