+++ services/multi_runtime/__init__.py
@@ -1,0 +1,7 @@
+"""Multi-Runtime Native Service.
+
+Multimodal audio/image/text inference daemon for multi_runtime.
+Replaces services.gemma4_audio_runtime as the primary implementation package.
+"""
+
+__version__ = "0.1.0"
